Add version info to instantiation and switch to mod.rs for file structure because it's way cleaner
