+++ Cargo.toml
@@ -45,9 +45,10 @@ optimize = """docker run --rm -v "$(pwd)":/code \
 provwasm-std = { version = "1.0.0-beta"}
 cosmwasm-std = { version = "1.0.0-beta" }
 cosmwasm-storage = { version = "1.0.0-beta" }
-cw-storage-plus = "0.8.0"
-cw2 = "0.8.1"
+cw-storage-plus = "0.12.1"
+cw2 = "0.12.1"
 schemars = "0.8.3"
+semver = "1"
 serde = { version = "1.0.127", default-features = false, features = ["derive"] }
 serde-json-wasm = { version = "0.3.1" }
 thiserror = { version = "1.0.26" }
