+++ src/execute/register_payable.rs
@@ -192,7 +192,7 @@ mod tests {
     use crate::core::error::ContractError;
     use crate::core::error::ContractError::Std;
     use crate::core::msg::ExecuteMsg;
-    use crate::test_utilities::testutils::{
+    use crate::testutil::test_utilities::{
         default_register_payable, get_duped_scope, test_instantiate, InstArgs,
         DEFAULT_FEE_COLLECTION_ADDRESS, DEFAULT_INFO_NAME, DEFAULT_ONBOARDING_DENOM,
         DEFAULT_PAYABLE_DENOM, DEFAULT_PAYABLE_TOTAL, DEFAULT_PAYABLE_TYPE, DEFAULT_PAYABLE_UUID,
