+++ src/instantiate/init_contract.rs
@@ -4,6 +4,7 @@ use crate::core::state::{config, State};
 use crate::util::conversions::to_percent;
 use cosmwasm_std::{Decimal, DepsMut, Env, MessageInfo, Response, Uint128};
 use provwasm_std::{bind_name, NameBinding, ProvenanceMsg, ProvenanceQuery};
+use crate::migrate::version_info::migrate_version_info;
 
 pub fn init_contract(
     deps: DepsMut<ProvenanceQuery>,
@@ -45,6 +46,9 @@ pub fn init_contract(
         NameBinding::Restricted,
     )?;
 
+    // Set the version info to the default contract values on instantiation
+    migrate_version_info(deps.storage)?;
+
     // Dispatch messages and emit event attributes
     Ok(Response::new()
         .add_message(bind_name_msg)
@@ -54,11 +58,12 @@ pub fn init_contract(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_utilities::testutils::{test_instantiate, InstArgs, DEFAULT_ONBOARDING_DENOM};
+    use crate::testutil::test_utilities::{test_instantiate, InstArgs, DEFAULT_ONBOARDING_DENOM};
     use cosmwasm_std::testing::mock_info;
     use cosmwasm_std::{coin, CosmosMsg, StdError};
     use provwasm_mocks::mock_dependencies;
     use provwasm_std::{NameMsgParams, ProvenanceMsgParams};
+    use crate::migrate::version_info::{CONTRACT_NAME, CONTRACT_VERSION, get_version_info};
 
     #[test]
     fn test_valid_init() {
@@ -122,6 +127,17 @@ mod tests {
             "oracle",
             generated_state.oracle_address.as_str(),
             "expected state to include the proper oracle address",
+        );
+        let version_info = get_version_info(deps.as_ref().storage).unwrap();
+        assert_eq!(
+            CONTRACT_NAME,
+            version_info.contract,
+            "the contract name should be properly stored after a successful instantiation",
+        );
+        assert_eq!(
+            CONTRACT_VERSION,
+            version_info.version,
+            "the contract version should be properly stored after a succesful instantiation",
         );
     }
