+++ src/lib.rs
@@ -4,9 +4,10 @@ pub mod contract;
 pub mod core;
 pub mod execute;
 pub mod instantiate;
+pub mod migrate;
 pub mod query;
 pub mod util;
 
 // Conditional modules
 #[cfg(feature = "enable-test-utils")]
-pub mod test_utilities;
+pub mod testutil;
