+++ src/migrate/mod.rs
@@ -1,0 +1,1 @@
+pub mod version_info;
