+++ src/testutil/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "enable-test-utils")]
+pub mod test_utilities;
